pyproject.toml
[project]
name = "pyvault-agent"
version = "0.1.0"
description = "Python implementation of Vault Agent with client-side caching and automatic authentication"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "hvac>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
